refactor/jni: move JNI utils to ffi_utils crate
