+++ ffi_utils/Cargo.toml
@@ -12,6 +12,7 @@ version = "0.5.0"
 [dependencies]
 clap = "=2.25.1"
 base64 = "~0.9.0"
+jni = "~0.10.1"
 log = "~0.4.1"
 moz-cheddar = "~0.4.2"
 regex = "~0.2.5"
