+++ ffi_utils/src/lib.rs
@@ -42,6 +42,7 @@
 #![cfg_attr(feature="cargo-clippy", allow(implicit_hasher, too_many_arguments, use_debug))]
 
 extern crate base64;
+extern crate jni;
 #[macro_use]
 extern crate log;
 #[macro_use]
@@ -62,6 +63,7 @@ pub mod test_utils;
 pub mod string;
 pub mod header_gen;
 pub mod bindgen_utils;
+pub mod java;
 
 pub use self::b64::{base64_decode, base64_encode};
 pub use self::catch_unwind::catch_unwind_cb;
